+++ src/pages/index.tsx
@@ -1,13 +1,47 @@
+import { Dropdown, Menu } from 'antd';
 import { NextPage } from 'next';
 
 const Home: NextPage = () => {
-  const publicKey = process.env.NEXT_PUBLIC_PUBLIC_KEY;
-  const privateKey = process.env.PRIVATE_KEY;
+  const menu = (
+    <Menu mode="vertical">
+      <Menu.Item>
+        <a
+          target="_blank"
+          rel="noopener noreferrer"
+          href="https://www.antgroup.com"
+        >
+          1st menu item
+        </a>
+      </Menu.Item>
+      <Menu.Item disabled>
+        <a
+          target="_blank"
+          rel="noopener noreferrer"
+          href="https://www.aliyun.com"
+        >
+          2nd menu item (disabled)
+        </a>
+      </Menu.Item>
+      <Menu.Item disabled>
+        <a
+          target="_blank"
+          rel="noopener noreferrer"
+          href="https://www.luohanacademy.com"
+        >
+          3rd menu item (disabled)
+        </a>
+      </Menu.Item>
+      <Menu.Item danger>a danger item</Menu.Item>
+    </Menu>
+  );
   return (
     <div className="w-full flex flex-col space-y-4 justify-center items-center">
-      <h1>Navegue pelas páginas através do cabeçalho</h1>
-      {publicKey}
-      {privateKey}
+      <Dropdown overlay={menu}>
+        <a className="ant-dropdown-link" onClick={e => e.preventDefault()}>
+          Hover me
+        </a>
+      </Dropdown>
+      ,
     </div>
   );
 };